requirements.txt
# This file lists all the dependencies required to run AutoNFT

# Core libraries
Pillow>=8.0.0  # For image processing in ImageGenerator
web3>=5.17.0  # For interacting with Ethereum blockchain in EthereumMinter

# Additional utilities
requests>=2.25.1  # For making HTTP requests to NFT marketplaces if needed

# Development tools (optional)
pytest>=6.2.2  # For running tests, if any are included in the project
